add findById

+++ src/main/java/com/base/recruitment/exception/client/ClientServiceException.java
@@ -1,13 +1,13 @@
 package com.base.recruitment.exception.client;
 
 public class ClientServiceException extends RuntimeException {
-    private long codigo;
-    public ClientServiceException(String message, long codigo) {
+    private int codigo;
+    public ClientServiceException(String message, int codigo) {
         super(message);
         this.codigo = codigo;
     }
 
-    public long getCodigo() {
+    public int getCodigo() {
         return this.codigo;
     }
 }
